refactor: fix the path ps input-related types
